user side coding start

+++ shop/views.py
@@ -12,8 +12,6 @@ from shop.forms import AddShopDetails ,AddItem
 import os
 import datetime
 
-
-
 def index(request):
     return render(request , 'index.html')
 
@@ -35,6 +33,7 @@ def loginshop(request):
                 'shopname':shopname,
                 'shopeamil':email,
                 'shop_id' : shop_id,
+                # 'id': 2,
             }
             # shop = authenticate(username=email, password=password)
             shop  = User.objects.get(email = email)
@@ -120,17 +119,24 @@ def AddShopData(request):
         if model.Shop.objects.filter(user = shop_id).exists():
             mutable_querydict = form.data.copy()
             mutable_querydict['user'] =shop_id
-            uploaded_file = request.FILES['shop_image']
-            date = datetime.date.today()
-            new_filename = str(shop_id)+'_shop_image'+str(date)+'.jpg'
+            nf= False
+            try:
+                uploaded_file = request.FILES['shop_image']
+                date = datetime.date.today()
+                new_filename = str(shop_id)+'_shop_image'+str(date)+'.jpg'
+                if uploaded_file:
+                    nf = True
+            except:
+                pass
             form.data = mutable_querydict.copy()
             print(form.data)
         if form.is_valid():
             shop_instance = form.save(commit=False)
-            shop_instance.shop_image.save(new_filename, uploaded_file)
+            if nf:  
+                shop_instance.shop_image.save(new_filename, uploaded_file)
+                if oldimg:
+                    os.remove(oldImgURL)
             shop_instance.save()
-            if oldimg:
-                os.remove(oldImgURL)
             return redirect('home')
     else :
         form = AddShopDetails(instance= shop)
@@ -141,12 +147,25 @@ def AddShopData(request):
     
 @login_required(login_url="loginshop")
 def Additem(request):
+    shop = request.session['shopdata']['shop_id']
+    shop_id = model.Shop.objects.get(user = shop).id
     if request.method == 'POST':
         form = AddItem(request.POST , request.FILES)
-        form.save()
+        mutable_querydict = form.data.copy()
+        mutable_querydict['shop_id'] =str(shop_id)
+        uploaded_file = request.FILES['item_image']
+        date = datetime.date.today()
+        new_filename = str(shop_id)+'_item_image_'+str(date)+'.jpg'
+        form.data = mutable_querydict.copy()
+        print(form.data)
+        if form.is_valid():
+            shop_instance = form.save(commit=False)
+            shop_instance.item_image.save(new_filename, uploaded_file)
+            shop_instance.save()
+            form.save()
     
     form = AddItem()
-    items = model.Item.objects.all().order_by("id")
+    items = model.Item.objects.filter(shop_id = shop_id)
     data={"items":items,"form":form}
 
     return render(request , 'additem.html', context = data)
